refactor: create form demo2 screen
